requirements/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
easyocr==1.6.2
entrypoints==0.4
fastapi==0.92.0
ffmpy==0.3.0
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.20.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio==2.26.0
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==15.0.6.1
lightning-utilities==0.7.1
linkify-it-py==2.0.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
networkx==3.0
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
orjson==3.8.7
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycryptodome==3.17
pydantic==1.10.5
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.9.4
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tifffile==2023.2.28
tks==0.2.1
tktimer==0.1.0
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.3
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
